requirements.txt
streamlit
anthropic
pyperclip
ftfy
trafilatura
newspaper4k
lxml
